+++ packages/editor/conversation/src/components/Comment.tsx
@@ -76,12 +76,13 @@ export default class Comment extends React.Component<Props, State> {
 
   shouldComponentUpdate(nextProps: Props, nextState: State) {
     const { isEditing, isReplying } = this.state;
-    const { isHighlighted } = this.props;
+    const { isHighlighted, portal } = this.props;
 
     if (
       nextState.isEditing !== isEditing ||
       nextState.isReplying !== isReplying ||
-      nextProps.isHighlighted !== isHighlighted
+      nextProps.isHighlighted !== isHighlighted ||
+      nextProps.portal !== portal
     ) {
       return true;
     }
@@ -400,6 +401,7 @@ export default class Comment extends React.Component<Props, State> {
       onEditorOpen,
       onEditorChange,
       sendAnalyticsEvent,
+      portal,
     } = this.props;
 
     if (!comments || comments.length === 0) {
@@ -430,6 +432,7 @@ export default class Comment extends React.Component<Props, State> {
         containerId={containerId}
         disableScrollTo={disableScrollTo}
         sendAnalyticsEvent={sendAnalyticsEvent}
+        portal={portal}
       />
     ));
   }