+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/index.ts
@@ -1,7 +1,6 @@
 // We need to import fetch polyfill for media components
 import 'whatwg-fetch';
 import * as assert from 'assert';
-import { undo } from 'prosemirror-history';
 import { EditorView } from 'prosemirror-view';
 
 import { ProviderFactory } from '@atlaskit/editor-common';
@@ -93,7 +92,9 @@ describe('Media plugin', () => {
     dropzoneContainer: HTMLElement = document.body,
     extraPlugins: any[] = [],
   ) => {
-    createAnalyticsEvent = jest.fn().mockReturnValue({ fire() {} });
+    createAnalyticsEvent = jest.fn().mockReturnValue({
+      fire() {},
+    });
     return createEditor({
       doc,
       editorPlugins: [
@@ -173,7 +174,6 @@ describe('Media plugin', () => {
           {
             id: '1',
             fileMimeType: 'image/jpeg',
-            fileId: Promise.resolve('foo'),
             fileName: 'foo.jpg',
             fileSize: 100,
             dimensions: {
@@ -187,7 +187,6 @@ describe('Media plugin', () => {
           {
             id: '2',
             fileMimeType: 'image/png',
-            fileId: Promise.resolve('bar'),
             fileName: 'bar.png',
             fileSize: 200,
             dimensions: {
@@ -200,26 +199,13 @@ describe('Media plugin', () => {
         pluginState.insertFiles(foo);
         pluginState.insertFiles(bar);
 
-        pluginState.stateManager.updateState('1', {
-          ...foo[0],
-          status: 'preview',
-          publicId: 'foo',
-        });
-
-        pluginState.stateManager.updateState('2', {
-          ...bar[0],
-          status: 'preview',
-          publicId: 'bar',
-        });
-
         expect(editorView.state.doc).toEqualDocument(
           doc(
             mediaSingle({
               layout: 'center',
             })(
               media({
-                id: 'foo',
-                __key: '1',
+                id: '1',
                 type: 'file',
                 collection: testCollectionName,
                 __fileName: 'foo.jpg',
@@ -233,8 +219,7 @@ describe('Media plugin', () => {
               layout: 'center',
             })(
               media({
-                id: 'bar',
-                __key: '2',
+                id: '2',
                 type: 'file',
                 collection: testCollectionName,
                 __fileName: 'bar.png',
@@ -259,7 +244,6 @@ describe('Media plugin', () => {
             editorView,
             media({
               id: temporaryFileId,
-              __key: temporaryFileId,
               type: 'file',
               collection: testCollectionName,
               __fileMimeType: 'image/png',
@@ -270,7 +254,6 @@ describe('Media plugin', () => {
             editorView,
             media({
               id: temporaryFileId,
-              __key: temporaryFileId,
               type: 'file',
               collection: testCollectionName,
               __fileMimeType: 'image/png',
@@ -285,7 +268,6 @@ describe('Media plugin', () => {
                     mediaSingle({ layout: 'center' })(
                       media({
                         id: temporaryFileId,
-                        __key: temporaryFileId,
                         type: 'file',
                         collection: testCollectionName,
                         __fileMimeType: 'image/png',
@@ -294,7 +276,6 @@ describe('Media plugin', () => {
                     mediaSingle({ layout: 'center' })(
                       media({
                         id: temporaryFileId,
-                        __key: temporaryFileId,
                         type: 'file',
                         collection: testCollectionName,
                         __fileMimeType: 'image/png',
@@ -319,7 +300,6 @@ describe('Media plugin', () => {
             editorView,
             media({
               id: temporaryFileId,
-              __key: temporaryFileId,
               type: 'file',
               collection: testCollectionName,
               __fileMimeType: 'image/png',
@@ -335,7 +315,6 @@ describe('Media plugin', () => {
                     mediaSingle({ layout: 'center' })(
                       media({
                         id: temporaryFileId,
-                        __key: temporaryFileId,
                         type: 'file',
                         collection: testCollectionName,
                         __fileMimeType: 'image/png',
@@ -366,7 +345,6 @@ describe('Media plugin', () => {
             fileSize: 200,
             fileMimeType: 'pdf',
             dimensions: { width: 200, height: 200 },
-            fileId: Promise.resolve('lala'),
           },
         ];
 
@@ -377,7 +355,6 @@ describe('Media plugin', () => {
             fileSize: 200,
             fileMimeType: 'image/png',
             dimensions: { width: 200, height: 200 },
-            fileId: Promise.resolve('bar'),
           },
         ];
 
@@ -386,23 +363,17 @@ describe('Media plugin', () => {
 
         pluginState.stateManager.updateState(id1, {
           ...lala[0],
-          status: 'preview',
-          publicId: 'lala',
         });
 
         pluginState.stateManager.updateState(id2, {
           ...bar[0],
-
-          status: 'preview',
-          publicId: 'bar',
         });
 
         expect(editorView.state.doc).toEqualDocument(
           doc(
             mediaGroup(
               media({
-                id: 'lala',
-                __key: id1,
+                id: id1,
                 type: 'file',
                 __fileMimeType: 'pdf',
                 __fileSize: 200,
@@ -412,8 +383,7 @@ describe('Media plugin', () => {
             ),
             mediaSingle({ layout: 'center' })(
               media({
-                id: 'bar',
-                __key: id2,
+                id: id2,
                 __fileName: 'bar.png',
                 __fileSize: 200,
                 height: 200,
@@ -435,9 +405,9 @@ describe('Media plugin', () => {
         await mediaProvider;
 
         pluginState.insertFiles([
-          { id: 'foo', fileMimeType: 'pdf', fileId: Promise.resolve('id1') },
-          { id: 'bar', fileMimeType: 'pdf', fileId: Promise.resolve('id2') },
-          { id: 'foobar', fileMimeType: 'pdf', fileId: Promise.resolve('id3') },
+          { id: 'foo', fileMimeType: 'pdf' },
+          { id: 'bar', fileMimeType: 'pdf' },
+          { id: 'foobar', fileMimeType: 'pdf' },
         ]);
 
         expect(editorView.state.doc).toEqualDocument(
@@ -445,21 +415,18 @@ describe('Media plugin', () => {
             mediaGroup(
               media({
                 id: 'foo',
-                __key: 'foo',
                 type: 'file',
                 __fileMimeType: 'pdf',
                 collection: testCollectionName,
               })(),
               media({
                 id: 'bar',
-                __key: 'bar',
                 type: 'file',
                 __fileMimeType: 'pdf',
                 collection: testCollectionName,
               })(),
               media({
                 id: 'foobar',
-                __key: 'foobar',
                 type: 'file',
                 __fileMimeType: 'pdf',
                 collection: testCollectionName,
@@ -470,126 +437,6 @@ describe('Media plugin', () => {
         );
       });
     });
-  });
-
-  it('should swap temporary id with public id', async () => {
-    const { editorView, pluginState } = editor(doc(p(), p('{<>}')));
-
-    const tempFileId = `temporary:${randomId()}`;
-
-    // wait until mediaProvider has been set
-    const provider = await mediaProvider;
-    // wait until mediaProvider's uploadContext has been set
-    await provider.uploadContext;
-
-    const publicId = 'public-id';
-    const fileState = [
-      { id: tempFileId, fileId: Promise.resolve('hello'), fileMimeType: 'pdf' },
-    ];
-
-    pluginState.insertFiles(fileState);
-    expect(editorView.state.doc).toEqualDocument(
-      doc(
-        p(),
-        mediaGroup(
-          media({
-            id: tempFileId,
-            __key: tempFileId,
-            type: 'file',
-            __fileMimeType: 'pdf',
-            collection: testCollectionName,
-          })(),
-        ),
-        p(),
-      ),
-    );
-
-    pluginState.stateManager.updateState(tempFileId, {
-      ...fileState[0],
-      status: 'preview',
-      publicId,
-    });
-
-    expect(editorView.state.doc).toEqualDocument(
-      doc(
-        p(),
-        mediaGroup(
-          media({
-            id: publicId,
-            __key: tempFileId,
-            type: 'file',
-            __fileMimeType: 'pdf',
-            collection: testCollectionName,
-          })(),
-        ),
-        p(),
-      ),
-    );
-
-    pluginState.destroy();
-  });
-
-  it('should not revert to temporary media nodes after upload finished and we undo', async () => {
-    const { editorView, pluginState } = editor(doc(p(), p('{<>}')));
-    const collectionFromProvider = jest.spyOn(
-      pluginState,
-      'collectionFromProvider' as any,
-    );
-    collectionFromProvider.mockImplementation(() => testCollectionName);
-    const tempFileId = `temporary:${randomId()}`;
-
-    // wait until mediaProvider has been set
-    const provider = await mediaProvider;
-    // wait until mediaProvider's uploadContext has been set
-    await provider.uploadContext;
-
-    pluginState.insertFiles([
-      { id: tempFileId, fileId: Promise.resolve('id') },
-    ]);
-
-    expect(editorView.state.doc).toEqualDocument(
-      doc(
-        p(),
-        mediaGroup(
-          media({
-            id: tempFileId,
-            __key: tempFileId,
-            type: 'file',
-            collection: testCollectionName,
-          })(),
-        ),
-        p(),
-      ),
-    );
-
-    // mark the upload as finished, triggering replacement of media node
-    stateManager.updateState(tempFileId, {
-      publicId: tempFileId,
-      status: 'preview',
-    });
-
-    expect(editorView.state.doc).toEqualDocument(
-      doc(
-        p(),
-        mediaGroup(
-          media({
-            id: tempFileId,
-            __key: tempFileId,
-            type: 'file',
-            collection: testCollectionName,
-          })(),
-        ),
-        p(),
-      ),
-    );
-
-    // undo last change
-    expect(undo(editorView.state, editorView.dispatch)).toBe(true);
-
-    expect(editorView.state.doc).toEqualDocument(doc(p(), p()));
-    collectionFromProvider.mockRestore();
-
-    pluginState.destroy();
   });
 
   it('should set new pickers exactly when new media provider is set', async () => {
@@ -939,22 +786,20 @@ describe('Media plugin', () => {
 
     const spy = jest.spyOn(editorView, 'focus');
 
-    pluginState.insertFiles([{ id: 'foo', fileId: Promise.resolve('id') }]);
+    pluginState.insertFiles([{ id: 'foo' }]);
     expect(spy).toHaveBeenCalled();
 
-    pluginState.insertFiles([{ id: 'bar', fileId: Promise.resolve('id') }]);
+    pluginState.insertFiles([{ id: 'bar' }]);
     expect(editorView.state.doc).toEqualDocument(
       doc(
         mediaGroup(
           media({
             id: 'bar',
-            __key: 'bar',
             type: 'file',
             collection: testCollectionName,
           })(),
           media({
             id: 'foo',
-            __key: 'foo',
             type: 'file',
             collection: testCollectionName,
           })(),
@@ -978,8 +823,6 @@ describe('Media plugin', () => {
     pluginState.insertFiles([
       {
         id: temporaryFileId,
-        fileId: Promise.resolve('id'),
-        status: 'preview',
         fileName: 'foo.png',
         fileSize: 1234,
         fileMimeType: 'pdf',
@@ -991,7 +834,6 @@ describe('Media plugin', () => {
         mediaGroup(
           media({
             id: temporaryFileId,
-            __key: temporaryFileId,
             type: 'file',
             collection: testCollectionName,
             __fileName: 'foo.png',
@@ -1319,7 +1161,6 @@ describe('Media plugin', () => {
           mediaGroup(
             media({
               id: pdfFile.id,
-              __key: pdfFile.id,
               type: 'file',
               __fileMimeType: pdfFile.fileMimeType,
               __fileName: pdfFile.fileName,
@@ -1345,7 +1186,6 @@ describe('Media plugin', () => {
           mediaGroup(
             media({
               id: pdfFile.id,
-              __key: pdfFile.id,
               type: 'file',
               __fileMimeType: pdfFile.fileMimeType,
               __fileName: pdfFile.fileName,
@@ -1364,7 +1204,6 @@ describe('Media plugin', () => {
         mediaGroup(
           media({
             id: pdfFile.id,
-            __key: pdfFile.id,
             type: 'file',
             __fileMimeType: pdfFile.fileMimeType,
             __fileName: pdfFile.fileName,
@@ -1385,7 +1224,7 @@ describe('Media plugin', () => {
           mediaGroup(
             media({
               id: pdfFile.id,
-              __key: pdfFile.id,
+
               type: 'file',
               __fileMimeType: pdfFile.fileMimeType,
               __fileName: pdfFile.fileName,
@@ -1394,7 +1233,7 @@ describe('Media plugin', () => {
             })(),
             media({
               id: pdfFile.id,
-              __key: pdfFile.id,
+
               type: 'file',
               __fileMimeType: pdfFile.fileMimeType,
               __fileName: pdfFile.fileName,
@@ -1481,6 +1320,61 @@ describe('Media plugin', () => {
       });
     });
 
+    describe('when media editor sends back save action', () => {
+      let pluginState: MediaPluginState;
+      let editorView: EditorView;
+      let closeMediaEditorSpy: any;
+
+      beforeEach(async () => {
+        const _editor = editor(
+          doc(
+            mediaSingle({ layout: 'center' })(
+              media({
+                id: 'media',
+                type: 'file',
+                collection: testCollectionName,
+              })(),
+            ),
+          ),
+        );
+        editorView = _editor.editorView;
+        pluginState = _editor.pluginState;
+
+        // wait for media provider so we set the upload context
+        await pluginState.setMediaProvider(mediaProvider);
+
+        setNodeSelection(editorView, 0);
+        closeMediaEditorSpy = jest.spyOn(pluginState, 'closeMediaEditor');
+        pluginState.openMediaEditor();
+        const toolbar = mountWithIntl<SmartMediaEditor['props'], {}>(
+          renderSmartMediaEditor(pluginState)!,
+        );
+        const { onUploadStart } = toolbar.props();
+        onUploadStart(
+          { id: 'some-new-id', mediaItemType: 'file' },
+          { width: 200, height: 100 },
+        );
+      });
+      it('should close media editor', () => {
+        expect(closeMediaEditorSpy).toHaveBeenCalled();
+      });
+      it('should replace media in the content', () => {
+        expect(editorView.state.doc).toEqualDocument(
+          doc(
+            mediaSingle({ layout: 'center' })(
+              media({
+                id: 'some-new-id',
+                collection: testCollectionName,
+                type: 'file',
+                height: 100,
+                width: 200,
+              })(),
+            ),
+          ),
+        );
+      });
+    });
+
     it('replaces the editing media node with a new one', async () => {
       const { pluginState, editorView } = editor(
         doc(
@@ -1497,11 +1391,17 @@ describe('Media plugin', () => {
       await pluginState.setMediaProvider(mediaProvider);
       setNodeSelection(editorView, 0);
       pluginState.openMediaEditor();
-      pluginState.replaceEditingMedia({
-        id: 'media2',
-        collectionName: 'collection2',
-        mediaItemType: 'file',
-      });
+      pluginState.replaceEditingMedia(
+        {
+          id: 'media2',
+          collectionName: 'collection2',
+          mediaItemType: 'file',
+        },
+        {
+          height: 100,
+          width: 200,
+        },
+      );
 
       expect(editorView.state.doc).toEqualDocument(
         doc(
@@ -1510,6 +1410,8 @@ describe('Media plugin', () => {
               id: 'media2',
               type: 'file',
               collection: 'collection2',
+              height: 100,
+              width: 200,
             })(),
           ),
         ),
@@ -1539,11 +1441,17 @@ describe('Media plugin', () => {
       insertText(editorView, 'add', refs['<>']);
 
       // should replace the old one
-      pluginState.replaceEditingMedia({
-        id: 'media2',
-        collectionName: 'collection2',
-        mediaItemType: 'file',
-      });
+      pluginState.replaceEditingMedia(
+        {
+          id: 'media2',
+          collectionName: 'collection2',
+          mediaItemType: 'file',
+        },
+        {
+          height: 100,
+          width: 200,
+        },
+      );
 
       expect(editorView.state.doc).toEqualDocument(
         doc(
@@ -1552,6 +1460,8 @@ describe('Media plugin', () => {
               id: 'media2',
               type: 'file',
               collection: 'collection2',
+              height: 100,
+              width: 200,
             })(),
           ),
           p('hello {<>}addworld'),