+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/media-with-fake-timers.ts
@@ -64,7 +64,6 @@ describe('Media plugin', async () => {
         {
           id: 'foo',
           fileMimeType: 'image/jpeg',
-          fileId: Promise.resolve('id'),
         },
       ]);
       jest.runOnlyPendingTimers();
@@ -81,7 +80,6 @@ describe('Media plugin', async () => {
         {
           id: 'foo',
           fileMimeType: 'image/jpeg',
-          fileId: Promise.resolve('id'),
           status: 'preview',
         },
       ]);
@@ -109,7 +107,6 @@ describe('Media plugin', async () => {
         {
           id: 'foo',
           fileMimeType: 'image/jpeg',
-          fileId: Promise.resolve('id'),
           status: 'preview',
         },
       ]);
@@ -117,7 +114,6 @@ describe('Media plugin', async () => {
         {
           id: 'bar',
           fileMimeType: 'image/jpeg',
-          fileId: Promise.resolve('id2'),
           status: 'preview',
         },
       ]);