+++ packages/editor/editor-core/src/__tests__/visual-regression/mobile-dark-editor.ts
@@ -1,6 +1,7 @@
 import { initEditor, snapshot } from './_utils';
 
-describe('Snapshot Test: Mobile Dark Editor', () => {
+// https://product-fabric.atlassian.net/browse/ED-6435
+describe.skip('Snapshot Test: Mobile Dark Editor', () => {
   let page;
   beforeAll(async () => {
     // @ts-ignore
@@ -9,6 +10,6 @@ describe('Snapshot Test: Mobile Dark Editor', () => {
   });
 
   it('should correctly render dark mode in mobile editor', async () => {
-    await snapshot(page);
+    await snapshot(page, 0.2);
   });
 });