+++ packages/editor/editor-core/src/plugins/table/pm-plugins/table-resizing/actions.ts
@@ -3,6 +3,7 @@ import { EditorView } from 'prosemirror-view';
 import { Node as PMNode } from 'prosemirror-model';
 import { EditorState } from 'prosemirror-state';
 import { findDomRefAtPos } from 'prosemirror-utils';
+import { gridSize } from '@atlaskit/theme';
 import {
   tableCellMinWidth,
   akEditorTableNumberColumnWidth,
@@ -229,23 +230,38 @@ export function scaleTable(
   pos: number,
   containerWidth: number | undefined,
   initialScale?: boolean,
+  parentWidth?: number,
   dynamicTextSizing?: boolean,
 ) {
   if (!tableElem) {
     return;
   }
 
+  // If a table has not been resized yet, columns should be auto.
+  if (hasTableBeenResized(node) === false) {
+    // If its not a re-sized table, we still want to re-create cols
+    // To force reflow of columns upon delete.
+    recreateResizeColsByNode(tableElem, node);
+    return;
+  }
+
   const start = pos + 1;
-  const state = scale(
-    view,
-    tableElem,
-    node,
-    start,
-    prevNode,
-    containerWidth,
-    initialScale,
-    dynamicTextSizing,
-  );
+
+  let state;
+  if (parentWidth) {
+    state = scaleWithParent(view, tableElem, parentWidth, node, start);
+  } else {
+    state = scale(
+      view,
+      tableElem,
+      node,
+      start,
+      prevNode,
+      containerWidth,
+      initialScale,
+      dynamicTextSizing,
+    );
+  }
 
   if (state) {
     const tr = applyColumnWidths(view, state, node, start);
@@ -269,27 +285,12 @@ function scale(
   node: PMNode,
   start: number,
   prevNode: PMNode,
-  containerWidth: number | undefined,
+  containerWidth?: number,
   initialScale?: boolean,
   dynamicTextSizing?: boolean,
 ): ResizeState | undefined {
-  // If a table has not been resized yet, columns should be auto.
-  if (hasTableBeenResized(node) === false) {
-    // If its not a re-sized table, we still want to re-create cols
-    // To force reflow of columns upon delete.
-    recreateResizeColsByNode(tableElem, node);
-    return;
-  }
-
   const maxSize = getLayoutSize(node.attrs.layout, containerWidth);
-  const resizer = Resizer.fromDOM(view, tableElem, {
-    minWidth: tableCellMinWidth,
-    maxSize,
-    node,
-    start,
-  });
 
-  let newWidth = maxSize;
   let prevTableWidth = getTableWidth(prevNode);
   let previousMaxSize = tableLayoutToSize[prevNode.attrs.layout];
 
@@ -301,6 +302,8 @@ function scale(
     );
   }
 
+  let newWidth = maxSize;
+
   // Determine if table was overflow
   if (prevTableWidth > previousMaxSize) {
     const overflowScale = prevTableWidth / previousMaxSize;
@@ -311,7 +314,38 @@ function scale(
     newWidth -= akEditorTableNumberColumnWidth;
   }
 
+  const resizer = Resizer.fromDOM(view, tableElem, {
+    minWidth: tableCellMinWidth,
+    maxSize,
+    node,
+    start,
+  });
+
   return resizer.scale(newWidth);
+}
+
+function scaleWithParent(
+  view: EditorView,
+  tableElem: HTMLTableElement,
+  parentWidth: number,
+  tableNode: PMNode,
+  start: number,
+) {
+  const resizer = Resizer.fromDOM(view, tableElem, {
+    minWidth: tableCellMinWidth,
+    maxSize: parentWidth,
+    node: tableNode,
+    start,
+  });
+
+  // Need to acount for the padding of the extension.
+  parentWidth -= gridSize() * 4;
+
+  if (tableNode.attrs.isNumberColumnEnabled) {
+    parentWidth -= akEditorTableNumberColumnWidth;
+  }
+
+  return resizer.scale(Math.floor(parentWidth));
 }
 
 /**