+++ packages/elements/emoji/src/__tests__/visual-regression/picker.ts
@@ -17,11 +17,15 @@ describe('Snapshot Test', () => {
     // @ts-ignore
     page = global.page;
     await page.goto(url);
+    page.waitForSelector('div[data-emoji-picker-container="true"]');
   });
 
   it(`should render emoji picker`, async () => {
     const image = await takeScreenShot(page, url);
     // @ts-ignore
-    expect(image).toMatchProdImageSnapshot();
+    expect(image).toMatchProdImageSnapshot({
+      failureThreshold: '100',
+      failureThresholdType: 'pixel',
+    });
   });
 });