+++ packages/media/media-editor/src/react/smartMediaEditorLoader.tsx
@@ -12,7 +12,7 @@ interface AsyncSmartMediaEditorState {
 
 export default class AsyncSmartMediaEditor extends React.PureComponent<
   SmartMediaEditorProps & AsyncSmartMediaEditorState,
-  AsyncSmartMediaEditorState
+  AsyncSmartMediaEditorState & { isErrored: boolean }
 > {
   static displayName = 'AsyncSmartMediaEditor';
   static SmartMediaEditor?: typeof SmartMediaEditorType;
@@ -20,18 +20,30 @@ export default class AsyncSmartMediaEditor extends React.PureComponent<
   state = {
     // Set state value to equal to current static value of this class.
     SmartMediaEditor: AsyncSmartMediaEditor.SmartMediaEditor,
+    isErrored: false,
   };
 
   async componentWillMount() {
     if (!this.state.SmartMediaEditor) {
-      const module = await import(/* webpackChunkName:"@atlaskit-internal_smart-media-editor" */
-      './smartMediaEditor');
-      AsyncSmartMediaEditor.SmartMediaEditor = module.default;
-      this.setState({ SmartMediaEditor: module.default });
+      try {
+        const module = await import(/* webpackChunkName:"@atlaskit-internal_smart-media-editor" */
+        './smartMediaEditor');
+        AsyncSmartMediaEditor.SmartMediaEditor = module.default;
+        this.setState({ SmartMediaEditor: module.default });
+      } catch (e) {
+        // tslint:disable-next-line:no-console
+        console.error(e);
+        this.setState({ isErrored: true });
+      }
     }
   }
 
   render() {
+    const { isErrored } = this.state;
+
+    if (isErrored) {
+      return null;
+    }
     if (!this.state.SmartMediaEditor) {
       return (
         <ModalSpinner blankedColor={colors.N700A} invertSpinnerColor={true} />