+++ packages/navigation/atlassian-switcher/src/components/atlassian-switcher.tsx
@@ -1,9 +1,11 @@
 import * as React from 'react';
 import JiraSwitcher from './jira-switcher';
 import ConfluenceSwitcher from './confluence-switcher';
+import GenericSwitcher from './generic-switcher';
 import ErrorBoundary from './error-boundary';
 import IntlProvider from './intl-provider';
 import messages from '../utils/messages';
+import { Product } from '../types';
 
 interface AtlassianSwitcherProps {
   product: string;
@@ -21,11 +23,17 @@ const AtlassianSwitcher = ({
 }: AtlassianSwitcherProps) => {
   let Switcher: React.ReactType;
   switch (product) {
-    case 'jira':
+    case Product.JIRA:
       Switcher = JiraSwitcher;
       break;
-    case 'confluence':
+    case Product.CONFLUENCE:
       Switcher = ConfluenceSwitcher;
+      break;
+    case Product.HOME:
+    case Product.PEOPLE:
+    case Product.SITE_ADMIN:
+    case Product.TRUSTED_ADMIN:
+      Switcher = GenericSwitcher;
       break;
     default:
       if (process.env.NODE_ENV !== 'production') {