+++ packages/navigation/atlassian-switcher/src/types.ts
@@ -15,3 +15,12 @@ export interface CustomLink {
   label: string;
   link: string;
 }
+
+export enum Product {
+  CONFLUENCE = 'confluence',
+  HOME = 'home',
+  JIRA = 'jira',
+  PEOPLE = 'people',
+  SITE_ADMIN = 'site-admin',
+  TRUSTED_ADMIN = 'trusted-admin',
+}