+++ packages/core/global-navigation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/global-navigation
 
+## 5.7.6
+- Updated dependencies [6ee7b60c4a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/6ee7b60c4a):
+  - @atlaskit/atlassian-switcher@0.1.0
+
 ## 5.7.5
 - Updated dependencies [e7fa9e1308](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/e7fa9e1308):
   - @atlaskit/atlassian-switcher@0.0.9